Bugfixed to diagonalNetwork.js
Fixed 2 bugs:
1) Line 61:

"linear" networks (i.e. of type  a ----- b) weren't rendered correctly, due to a zero division. Inserted Math.max(1, ...) to fix this problem.

2) Line 43:

When re-rendering a network in the same parent, the svg viewBox, which ensures that all text labels are visible, was affected by the viewBox attribute before the re-rendering. Thus the viewBox attribute changed even if the content did not, and could even move the image out of bounds. By explicitly setting the viewBox attribute to null on initalization, this is now inhibited.